software/requirements.txt
Flask>=2.0.0
Flask-Cors>=3.0.0
google-generativeai>=0.1.0
spotipy>=2.0.0